// frontend/src/types/auth.ts
export interface User {
    user_id: string
    email: string
    username: string
    role: string
    created_at: string
}

export interface AuthResponse {
    access_token: string
    refresh_token: string
    token_type: string
    user: User
}

export interface LoginRequest {
    email: string
    password: string
}

export interface RegisterRequest {
    email: string
    password: string
    username: string
}